requirements.txt
tensorflow-cpu==2.15.0.post1
opencv-python==4.12.0.88
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.0
tqdm==4.66.1
Pillow==10.0.0
fpdf==1.7.2
python-docx==0.8.11